+++ CHANGELOG.md
@@ -19,4 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed a bug where it was not possible to read a singlet entry at height 0.
+
 - Fixed a bug when reading a single table row and it's present in the index, it was not picked up correctly.